+++ src/main/java/com/back/domain/post/postComment/controller/ApiV1PostCommentController.java
@@ -1,11 +1,12 @@
 package com.back.domain.post.postComment.controller;
 
 import com.back.domain.member.member.entity.Member;
-import com.back.domain.member.member.service.MemberService;
 import com.back.domain.post.post.entity.Post;
 import com.back.domain.post.post.service.PostService;
 import com.back.domain.post.postComment.dto.PostCommentDto;
 import com.back.domain.post.postComment.entity.PostComment;
+import com.back.global.exception.ServiceException;
+import com.back.global.rq.Rq;
 import com.back.global.rsData.RsData;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -24,7 +25,7 @@ import java.util.List;
 @Tag(name = "ApiV1PostCommentController", description = "API 댓글 컨트롤러")
 public class ApiV1PostCommentController {
     private final PostService postService;
-    private final MemberService memberService;
+    private final Rq rq;
 
     @GetMapping
     @Transactional(readOnly = true)
@@ -62,9 +63,14 @@ public class ApiV1PostCommentController {
             @PathVariable int postId,
             @PathVariable int id
     ) {
+        Member actor = rq.getActor();
+
         Post post = postService.findById(postId).get();
 
         PostComment postComment = post.findCommentById(id).get();
+
+        if (!actor.equals(postComment.getAuthor()))
+            throw new ServiceException("403-1", "댓글 삭제 권한이 없습니다.");
 
         postService.deleteComment(post, postComment);
 
@@ -90,9 +96,14 @@ public class ApiV1PostCommentController {
             @PathVariable int id,
             @Valid @RequestBody PostCommentModifyReqBody reqBody
     ) {
+        Member actor = rq.getActor();
+
         Post post = postService.findById(postId).get();
 
         PostComment postComment = post.findCommentById(id).get();
+
+        if (!actor.equals(postComment.getAuthor()))
+            throw new ServiceException("403-1", "댓글 수정 권한이 없습니다.");
 
         postService.modifyComment(postComment, reqBody.content);
 
@@ -117,7 +128,7 @@ public class ApiV1PostCommentController {
             @PathVariable int postId,
             @Valid @RequestBody PostCommentWriteReqBody reqBody
     ) {
-        Member actor = memberService.findByUsername("user1").get(); // 임시로 작성자를 user1로 지정
+        Member actor = rq.getActor();
         Post post = postService.findById(postId).get();
 
         PostComment postComment = postService.writeComment(actor, post, reqBody.content);
